Additional 64-bit compatibility fixes, and enabling of write_range and read_range to broadcast addresses.

+++ applications/dsp_cmd/dsp_cmd.c
@@ -139,7 +139,7 @@ int main(int argc, char **argv)
 
 	while (!done) {
 
-		unsigned int n = LINE_LEN;
+		size_t n = LINE_LEN;
 
 		if ( options.cmd_now ) {
 			strcpy(line, options.cmd_command);
