+++ applications/example/example_data.c
@@ -24,7 +24,7 @@ typedef struct {
 } my_counter_t;
 
 
-int frame_callback(unsigned user_data, int size, u32 *data)
+int frame_callback(unsigned long user_data, int size, u32 *data)
 {
 	//Re-type 
 	my_counter_t *c = (my_counter_t*)user_data;
@@ -88,7 +88,7 @@ int main()
 	// each received frame.  (Other options here are to set up an
 	// output file, or a file sequence.)
 	mcedata_storage_t* ramb;
-	ramb = mcedata_rambuff_create( frame_callback, (unsigned)&counter );
+	ramb = mcedata_rambuff_create( frame_callback, (unsigned long)&counter );
 
 	// Setup an acqusition structure, associated with the rambuff.
 	// We can specify the number of readout rows, or pass -1 to use the
